don't kill and emit slaveDied() again if already marked as dead. If clients show a dialog e.g. after error/slaveDied is emitted, the timeout might occur while the dialog is shown, causing slaveDied() to be emitted twice. That results in an extra deref(), which then causes the slave to be deleted while the dialog is shown. Return from dialog, local deref() (e.g. in gotInput() ) => crash Might fix 191589 CCBUG:191589
svn path=/branches/KDE/4.3/kdelibs/; revision=1056317

+++ kio/kio/slave.cpp
@@ -125,9 +125,10 @@ void Slave::accept()
 void Slave::timeout()
 {
     Q_D(Slave);
+   if (d->dead)
+      return;
    if (d->connection->isConnected())
       return;
-
    kDebug(7002) << "slave failed to connect to application pid=" << d->m_pid << " protocol=" << d->m_protocol;
    if (d->m_pid && (::kill(d->m_pid, 0) == 0))
    {
